sort out ownership issues

+++ src/pref/config/config_definition.rs
@@ -104,7 +104,7 @@ pub struct SysMgmt {
 	#[serde(alias = "inhibitOnBehalf")]
 	pub conduct_inhibit:	bool,
 
-	pub uclamp_max:		u16,
+	pub uclamp_max:		u8,
 
 	#[serde(alias = "deviceAllow")]
 	#[serde(deserialize_with = "deserialise_device_allow")]
